REFACTOR: module rename for code style
