+++ src/utils/string-helpers.ts
